temporary front end with login component to test be functionality
