+++ package.json
@@ -10,14 +10,21 @@
   },
   "dependencies": {
     "@expo/metro-runtime": "~3.1.3",
+    "@react-navigation/native": "^6.1.15",
+    "@react-navigation/native-stack": "^6.9.24",
+    "axios": "^1.6.7",
     "expo": "~50.0.11",
     "expo-status-bar": "~1.11.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "0.73.4",
+    "react-native-image-picker": "^7.1.0",
     "react-native-view-shot": "^3.8.0",
     "react-native-web": "~0.19.6",
-    "tesseract.js": "^5.0.5"
+    "react-router-native": "^6.22.3",
+    "tesseract.js": "^5.0.5",
+    "react-native-screens": "~3.29.0",
+    "react-native-safe-area-context": "4.8.2"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
